Improved breakpoint handling, by persisting Source to marker attributes, when creating breakpoints Breakpoints are now passed to DebugLaunchConfiguration CommandMessage

+++ src/monto/eclipse/launching/debug/MontoToggleBreakpointsTarget.java
@@ -14,6 +14,7 @@ import org.eclipse.jface.viewers.ISelection;
 import org.eclipse.ui.IWorkbenchPart;
 
 import monto.eclipse.Activator;
+import monto.eclipse.MontoParseController;
 
 public class MontoToggleBreakpointsTarget implements IToggleBreakpointsTarget {
 
@@ -22,27 +23,33 @@ public class MontoToggleBreakpointsTarget implements IToggleBreakpointsTarget {
       throws CoreException {
     System.out.println("MontoToggleBreakpointsTarget.toggleLineBreakpoints()");
     if (part != null && part instanceof UniversalEditor) {
-      UniversalEditor textEditor = (UniversalEditor) part;
-      IResource resource = (IResource) textEditor.getEditorInput().getAdapter(IResource.class);
-      ITextSelection textSelection = (ITextSelection) selection;
-      int lineNumber = textSelection.getStartLine();
+      UniversalEditor universalEditor = (UniversalEditor) part;
+      if (universalEditor.getParseController() instanceof MontoParseController) {
+        MontoParseController montoParseController =
+            (MontoParseController) universalEditor.getParseController();
 
-      // delete old breakpoints in lineNumber
-      IBreakpoint[] oldBreakpoints =
-          DebugPlugin.getDefault().getBreakpointManager().getBreakpoints(Activator.PLUGIN_ID);
-      boolean deletedBreakPoints = false;
-      for (IBreakpoint oldBreakpoint : oldBreakpoints) {
-        if (resource.equals(oldBreakpoint.getMarker().getResource())
-            && ((ILineBreakpoint) oldBreakpoint).getLineNumber() == (lineNumber + 1)) {
-          deletedBreakPoints = true;
-          oldBreakpoint.delete();
+        IResource resource =
+            (IResource) universalEditor.getEditorInput().getAdapter(IResource.class);
+        ITextSelection textSelection = (ITextSelection) selection;
+        int lineNumber = textSelection.getStartLine();
+
+        // delete old breakpoints in lineNumber
+        IBreakpoint[] oldBreakpoints =
+            DebugPlugin.getDefault().getBreakpointManager().getBreakpoints(Activator.PLUGIN_ID);
+        boolean deletedBreakPoints = false;
+        for (IBreakpoint oldBreakpoint : oldBreakpoints) {
+          if (resource.equals(oldBreakpoint.getMarker().getResource())
+              && ((ILineBreakpoint) oldBreakpoint).getLineNumber() == (lineNumber + 1)) {
+            deletedBreakPoints = true;
+            oldBreakpoint.delete();
+          }
         }
-      }
 
-      if (!deletedBreakPoints) {
-        // create new one
-        DebugPlugin.getDefault().getBreakpointManager()
-            .addBreakpoint(new MontoLineBreakpoint(resource, lineNumber + 1));
+        if (!deletedBreakPoints) {
+          // create new one
+          DebugPlugin.getDefault().getBreakpointManager().addBreakpoint(
+              new MontoLineBreakpoint(resource, lineNumber + 1, montoParseController.getSource()));
+        }
       }
     }
   }
